Rename rapp reducer
