+++ src/index.jsx
@@ -1,20 +1,14 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import { compose, createStore, applyMiddleware } from "redux";
+
 import { Provider } from "react-redux";
-import thunk from "redux-thunk";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
+import { store } from "./services/store";
 import App from "./components/app/app";
 import reportWebVitals from "./reportWebVitals";
-import { rootReducer } from "./services/reducers/index";
+
 import "./index.css";
-
-const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-
-const enhancer = composeEnhancers(applyMiddleware(thunk));
-
-const store = createStore(rootReducer, enhancer);
 
 ReactDOM.render(
   <React.StrictMode>
