+++ src/services/store.js
@@ -1,0 +1,8 @@
+import { compose, createStore, applyMiddleware } from "redux";
+import thunk from "redux-thunk";
+import { rootReducer } from "./reducers";
+
+const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
+const enhancer = composeEnhancers(applyMiddleware(thunk));
+
+export const store = createStore(rootReducer, enhancer);
